openapi: Fix validity issues
Some parameters were wrongly specified as path parameters, while they actually were query parameters. Additionally, the don't have to be specified in the openapi path itself.

The schema of the requestBody for posting a new gallery image was missing - but it does have to be set to binary string. See https://swagger.io/docs/specification/describing-responses/ (response that returns a file).

The gallery array for a project version was missing the item type. I set it to string because it is an array of URLs.